Fix some vulnerabilities
