+++ frontend/Dockerfile
@@ -1,20 +1,19 @@
-FROM node:carbon
+FROM node:fermium
 
 RUN mkdir -p /usr/src/app
 WORKDIR /usr/src/app
 
-RUN apt-get update
+RUN apt-get update && apt-get upgrade -y
+RUN apt-get -y autoremove
 RUN apt-get install -y libtool pkg-config build-essential autoconf automake
-RUN apt-get install -y libzmq3-dev
 
 COPY package.json /usr/src/app/
 RUN npm install
-
 RUN npm install -g envify browserify
 
 # Bundle app source
 COPY . /usr/src/app
 
-EXPOSE 3000 3001 6080
-
-CMD echo $MAPBOX_TOKEN && npm run build && npm start
+EXPOSE 3000 6080
+CMD ["/bin/sh"]
+# CMD npm run build && npm start
