+++ frontend/bin/www
@@ -6,7 +6,10 @@
 
 var app = require('../app');
 var debug = require('debug')('latency:server');
-var http = require('http');
+
+
+
+
 
 /**
  * Get port from environment and store in Express.
@@ -19,7 +22,7 @@ app.set('port', port);
  * Create HTTP server.
  */
 
-var server = http.createServer(app);
+const server = app.get('srv');
 
 /**
  * Listen on provided port, on all network interfaces.
