+++ frontend/package.json
@@ -3,35 +3,34 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "build": "NODE_ENV=production browserify src/react.js -o public/javascripts/bundle.js -t babelify  -t sassify -t glslify -t envify",
+    "build": "NODE_ENV=production browserify src/react.js -o public/javascripts/bundle.js -t babelify  -t sassify -t glslify -t envify --debug",
     "start": "node bin/www"
   },
   "dependencies": {
-    "autobind-decorator": "^1.3.3",
     "babel-polyfill": "^6.1.19",
-    "body-parser": "~1.15.2",
+    "body-parser": "^1.19.0",
     "cookie-parser": "~1.4.3",
-    "debug": "~2.2.0",
+    "debug": "^4.3.2",
     "deck.gl": "^3.0.7",
     "earcut": "^2.0.6",
-    "express": "~4.14.0",
+    "express": "^4.17.1",
     "extrude-polyline": "^1.0.6",
     "geojson-normalize": "0.0.1",
     "get-pixels": "^3.3.0",
     "gl-matrix": "^2.3.2",
-    "glslify": "^5.0.0",
+    "glslify": "^7.1.1",
     "influx": "^5.0.0-alpha.4",
-    "jade": "~1.11.0",
     "lodash.flattendeep": "^3.0.2",
     "luma.gl": "^2.10.0",
-    "morgan": "~1.7.0",
+    "morgan": "^1.10.0",
+    "pug": "^3.0.2",
     "react-map-gl": "^1.7.2",
     "save-pixels": "^2.3.0",
-    "serve-favicon": "~2.3.0",
-    "socket.io": "~1.6.0",
+    "serve-favicon": "^2.5.0",
+    "socket.io": "^4.1.2",
     "viewport-mercator-project": "3.0.0-beta2",
     "webgl-obj-loader": "^0.1.1",
-    "zmq": "~2.15.3"
+    "zeromq": "^5.2.8"
   },
   "devDependencies": {
     "babel-cli": "^6.3.15",
@@ -48,14 +47,14 @@
     "babelify": "^7.2.0",
     "brfs-babel": "^1.0.0",
     "browserify": "13.0.0",
-    "budo": "^8.0.3",
+    "budo": "^11.6.4",
     "d3-color": "^1.0.1",
     "d3-request": "^1.0.2",
     "d3-scale": "^1.0.3",
     "discify": "^1.2.0",
     "electron-prebuilt": "^0.37.2",
     "envify": "^3.4.0",
-    "eslint": "^2.0.0",
+    "eslint": "^2.13.1",
     "eslint-config-uber-es2015": "^2.0.0",
     "eslint-config-uber-jsx": "^2.0.0",
     "eslint-plugin-react": "^5.0.0",
@@ -63,8 +62,7 @@
     "gl": "^4.0.1",
     "highlight.js": "^9.7.0",
     "immutable": "^3.7.5",
-    "luma.gl": "^2.10.0",
-    "marked": "^0.3.6",
+    "marked": "^2.1.3",
     "react": "^15.3.2",
     "react-addons-pure-render-mixin": "^15.3.2",
     "react-addons-test-utils": "^15.0.2",
@@ -74,7 +72,7 @@
     "react-router": "^2.8.1",
     "redux": "^3.6.0",
     "redux-thunk": "^2.1.0",
-    "sassify": "^2.0.0",
+    "sassify": "^4.0.1",
     "stats.js": "^0.17.0",
     "tap-browser-color": "^0.1.2",
     "tape": "^4.6.0",
