+++ frontend/views/index.pug
