+++ frontend/views/layout.pug
@@ -3,9 +3,9 @@ html
   head
     title= title
     link(rel='stylesheet', href='//fonts.googleapis.com/css?family=Roboto')
-    script(src="//cdnjs.cloudflare.com/ajax/libs/socket.io/1.7.3/socket.io.js")
+    script(src="//cdnjs.cloudflare.com/ajax/libs/socket.io/4.1.2/socket.io.js")
     script.
-      var socket = io.connect(':3001');
+      var socket = io.connect('localhost:3000');
     script(src='//ajax.googleapis.com/ajax/libs/jquery/3.1.0/jquery.min.js')
     script(src='//cdnjs.cloudflare.com/ajax/libs/tether/1.3.7/js/tether.min.js')
     link(rel='stylesheet', href='//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css')
