+++ frontend/views/stats.pug
